feat: change int64 to int32

+++ internal/dto/roles/get.go
@@ -1,11 +1,11 @@
 package dto
 
 type GetRoleReq struct {
-	Id int64 `json:"id" validate:"required"`
+	Id int32 `json:"id" validate:"required"`
 }
 
 type GetRoleRes struct {
-	Id          int64  `json:"id"`
+	Id          int32  `json:"id"`
 	RoleName    string `json:"roleName"`
 	Description string `json:"description"`
 }
